Add troubleshooting tips for EACCESS error
Turns out that this permission denied error can come up in many different
cases, not just FreeBSD or Docker, but also Ubuntu. Also seems that there
are at least two solutions, one of which may or may not resolve the issue,
but has worked for some. The second solution almost certainly works, but has
some drawback in terms of security implications.

Relates to BuckleScript/bucklescript issues 2051 and 3028.